+++ packages/terra-framework-docs/src/terra-dev-site/doc/table/Examples/TableSingleRowSelection.jsx
@@ -63,7 +63,9 @@ const TableSingleRowSelection = () => {
   const { cols, rows } = tableDataJSON;
   const [rowData, setRowData] = useState(rows);
 
-  const onRowSelect = useCallback((rowId) => {
+  const onRowSelect = useCallback((rowSelection) => {
+    const { rowId } = rowSelection;
+
     const newRowData = [...rowData];
 
     const dataRowToUpdate = newRowData.find(row => row.id === rowId);