+++ packages/terra-table/src/Table.jsx
@@ -291,7 +291,7 @@ function Table(props) {
 
   const handleCellSelection = useCallback((selectionDetails) => {
     if (!isGridContext && onRowSelect) {
-      onRowSelect(selectionDetails.rowId);
+      onRowSelect({ sectionId: selectionDetails.sectionId, rowId: selectionDetails.rowId });
       return;
     }