+++ packages/terra-framework-docs/CHANGELOG.md
@@ -5,6 +5,9 @@
 * Added
   * Added documentation for exported constants of `terra-table` in the About page.
 
+* Changed
+  * Updated examples and tests for `terra-table` to consume updated onRowSelect callback.
+
 ## 1.49.0 - (December 1, 2023)
 
 * Added